move type-check to core
